+++ read_structure_step/write.py
@@ -56,17 +56,12 @@ def write(
     """
 
     if type(file_name) is not str:
-        raise TypeError(
-            """write_structure_step: The file name must be a string, but a
-            %s was given. """
-            % str(type(file_name))
-        )
+        raise TypeError("""write_structure_step: The file name must be a string, but a
+            %s was given. """ % str(type(file_name)))
 
     if file_name == "":
-        raise NameError(
-            """write_structure_step: The file name for the structure file
-            was not specified."""
-        )
+        raise NameError("""write_structure_step: The file name for the structure file
+            was not specified.""")
 
     file_name = os.path.abspath(file_name)